moved file info update to dev tools file/url group
