+++ pages/urls.py
@@ -63,11 +63,11 @@ urlpatterns = [
     path('api-geteml/<str:emlName>', views.getEml, name='api-geteml'),
     path('api-removeCentcom/<uuid:id>', views.removeCentcom, name='api-removeCentcom'),
     path('api-requestnotes/<uuid:requestid>', views.requestNotes, name='api-requestnotes'),
-    path('api-updateFileInfo', views.updateFiles, name='api-updateFileInfo'),
 
     # dev tools
     path('tools-makefiles', views.makeFiles, name='make-files'),
     path('tools-stubget', views.stubGet, name='stub-get'),
     path('tools-stubpost', views.stubPost, name='stub-post'),
     path('tools-setupdb', views.setupDB, name="setupdb"),
+    path('tools-updateFileInfo', views.updateFiles, name='api-updateFileInfo'),
 ]
