+++ pages/views/__init__.py
@@ -8,4 +8,3 @@ from .dev_tools import *
 from .reports import *
 from .resources import *
 from .feedback import *
-from .file_info import *
